requirements.txt
pyspark==3.4.4
delta-spark==2.4.0
py4j
numpy
pandas